+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.11 (2022-10-25)
+
+- Fixed Vault backend caching
+
+## 1.0.10 (2022-10-25)
+
+- Fixed Vault backend caching
+
+## 1.0.9 (2022-10-25)
+
+- Fixed Vault backend caching
+
+## 1.0.8 (2022-10-25)
+
+- Updated requirements to replace `cachetools` by `cacheout` 
+
 ## 1.0.7 (2022-10-24)
 
 - Updated requirement cachetools<5.0.0